api-server: refine typing codes
