+++ packages/api-server/src/plugins/cache/cache-builder.ts
@@ -1,6 +1,4 @@
 import CacheBuilder from "../../core/cache/CacheBuilder";
-import { FastifyJWTOptions } from "@fastify/jwt";
-import { FastifyOAuth2Options } from "@fastify/oauth2";
 import fp from "fastify-plugin";
 import pino from "pino";
 import {ConnectionWrapper} from "../../core/db/connection-wrapper";
@@ -11,7 +9,7 @@ declare module 'fastify' {
     }
 }
 
-export default fp<FastifyOAuth2Options & FastifyJWTOptions>(async (fastify) => {
+export default fp(async (fastify) => {
     const log = fastify.log as pino.Logger;
     const wrapper = await ConnectionWrapper.new({
         uri: fastify.config.DATABASE_URL
