+++ packages/api-server/src/plugins/executors/tidb-playground-query-executor.ts
@@ -1,10 +1,8 @@
-import { FastifyJWTOptions } from '@fastify/jwt';
-import { FastifyOAuth2Options } from '@fastify/oauth2';
 import fp from 'fastify-plugin';
 import { TiDBPlaygroundQueryExecutor } from '../../core/executor/query-executor/TiDBPlaygroundQueryExecutor';
 import { getPlaygroundSessionLimits } from '../../core/playground/limitation';
 
-export default fp<FastifyOAuth2Options & FastifyJWTOptions>(async (app) => {
+export default fp(async (app) => {
   app.decorate('playgroundQueryExecutor', new TiDBPlaygroundQueryExecutor({
     uri: app.config.PLAYGROUND_DATABASE_URL,
   }, getPlaygroundSessionLimits()));
