+++ packages/api-server/src/routes/explorer/questions/_questionId/recommend/index.ts
@@ -1,5 +1,5 @@
 import {FastifyPluginAsyncJsonSchemaToTs} from "@fastify/type-provider-json-schema-to-ts";
-import {Auth0User, parseAuth0User} from "../../../../../plugins/services/user-service/auth0";
+import {Auth0User, parseAuth0User} from "../../../../../plugins/auth/auth0";
 
 const recommendQuestionSchema = {
   summary: 'Recommend question',
@@ -21,7 +21,6 @@ export const recommendQuestionHandler: FastifyPluginAsyncJsonSchemaToTs = async 
     Params: recommendQuestionParams;
   }>('/', {
     schema: recommendQuestionSchema,
-    // @ts-ignore
     preValidation: app.authenticate
   }, async function (req, reply) {
     const { questionId } = req.params;
