+++ packages/api-server/src/routes/repos/_owner/_repo/subscribe/index.ts
@@ -1,5 +1,5 @@
 import {FastifyPluginAsyncJsonSchemaToTs} from "@fastify/type-provider-json-schema-to-ts";
-import { Auth0User, parseAuth0User } from "../../../../../plugins/services/user-service/auth0";
+import { Auth0User, parseAuth0User } from "../../../../../plugins/auth/auth0";
 
 export interface IParams {
     owner: string;
@@ -30,7 +30,6 @@ const root: FastifyPluginAsyncJsonSchemaToTs = async (app, opts): Promise<void> 
     app.put<{
         Params: IParams;
     }>('/', {
-        // @ts-ignore
         preValidation: app.authenticate,
         schema: schema
     }, async function (req, reply) {
