+++ packages/api-server/src/routes/user/index.ts
@@ -1,5 +1,5 @@
 import {FastifyPluginAsyncJsonSchemaToTs} from "@fastify/type-provider-json-schema-to-ts/index";
-import { Auth0User, parseAuth0User } from "../../plugins/services/user-service/auth0";
+import { Auth0User, parseAuth0User } from "../../plugins/auth/auth0";
 
 const schema = {
     description: 'Get the authenticated user',
@@ -46,7 +46,6 @@ const schema = {
 
 const root: FastifyPluginAsyncJsonSchemaToTs = async (app, opts): Promise<void> => {
     app.get('/', {
-        // @ts-ignore
         preValidation: app.authenticate,
         schema
     }, async function (req, reply) {
